% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_data.R
\docType{data}
\name{cds_data}
\alias{cds_data}
\title{Common Data Set Extracts}
\format{
A \code{data.frame} with a report in each row, and fields in each column.
A few columns contain metadata:
\describe{
  \item{school}{Name of the school.}
  \item{source}{URL of the original PDF file.}
  \item{md5}{MD5 hash of the original PDF file.}
  \item{retieved}{Date downloaded.}
  \item{processed}{Date processed.}
  \item{year}{Year of the report.}
}
The rest are named first by section and item number (e.g., B1), then by a combination
of levels such as the part, column name, and row name from the original table.
}
\usage{
cds_data
}
\description{
Extracted portions of school Common Data Set reports.
}
\examples{
cds_data[1:10, c(
  "school", "retrieved", "B1_men_full_new_undergrad", "B1_women_full_new_undergrad"
)]
}
\keyword{datasets}
