% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_get_reports.R
\name{cds_get_reports}
\alias{cds_get_reports}
\title{Download CDS Reports}
\usage{
cds_get_reports(school, out_dir, extract = TRUE, check = TRUE,
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{school}{Name of the school.}

\item{out_dir}{Path to an output directory, which should contain the output of
\code{\link{cds_get_links}} in a \code{record/{school}} sub-directory.}

\item{extract}{Logical; if \code{FALSE}, will not save text from the PDF file.}

\item{check}{Logical; if \code{FALSE}, will not check if the file is a Common Data Set report.}

\item{overwrite}{Logical; if \code{TRUE}, will always try to download all reports.}

\item{verbose}{Logical; if \code{FALSE}, will not show status messages.}
}
\value{
A list with an entry for each report, which is also saved as
\code{\{out_dir\}/record/\{school\}/manifest.json}:
\itemize{
  \item \strong{\code{url}}: Link to the report.
  \item \strong{\code{retrieved}}: Time the report was downloaded.
  \item \strong{\code{md5}}: MD5 sum of the report.
}
}
\description{
Download the CDS PDF reports from a given school,
using URLs retrieved with \code{\link{cds_get_links}}.
}
\examples{
if (dir.exists("../../reports")) {
  manifest <- cds_get_reports("Yale University", "../../reports")
  manifest[1:2]
}
}
