% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_process_report.R
\name{cds_process_report}
\alias{cds_process_report}
\title{Process a CDS Report}
\usage{
cds_process_report(school, file)
}
\arguments{
\item{school}{Name of the school, to be checked for in the document.}

\item{file}{Path to the PDF report.}
}
\value{
A \code{list} of data extracted from the report.
}
\description{
Extract data from a CDS PDF report.
}
\examples{
file <- "../../reports/text/Yale University/cds2008_2009_1.txt"
if (file.exists(file)) {
  values <- cds_process_report("Yale University", file)
  values[1:10]
}
}
