% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolCDS-package.R
\docType{package}
\name{schoolCDS-package}
\alias{schoolCDS}
\alias{schoolCDS-package}
\title{schoolCDS: School Common Data Set (CDS) Data}
\description{
Collect, process, and work with school Common Data Set (CDS) data (\url{https://commondataset.org}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://DISSC-yale.github.io/schoolCDS/}
  \item \url{https://github.com/DISSC-yale/schoolCDS}
  \item Report bugs at \url{https://github.com/DISSC-yale/schoolCDS/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Other contributors:
\itemize{
  \item Data-Intensive Social Science Center [funder, copyright holder]
}

}
\keyword{internal}
